# internal/lz4block: Faster table reset