--- a/internal/lz4block/block.go
+++ b/internal/lz4block/block.go
@@ -63,7 +63,10 @@ type Compressor struct {
 	// inspecting the input stream.
 	table [htSize]uint16
 
-	needsReset bool
+	// Bitmap indicating which positions in the table are in use.
+	// This allows us to quickly reset the table for reuse,
+	// without having to zero everything.
+	inUse [htSize / 32]uint32
 }
 
 // Get returns the position of a presumptive match for the hash h.
@@ -71,7 +74,10 @@ type Compressor struct {
 // If si < winSize, the return value may be negative.
 func (c *Compressor) get(h uint32, si int) int {
 	h &= htSize - 1
-	i := int(c.table[h])
+	i := 0
+	if c.inUse[h/32]&(1<<(h%32)) != 0 {
+		i = int(c.table[h])
+	}
 	i += si &^ winMask
 	if i >= si {
 		// Try previous 64kiB block (negative when in first block).
@@ -83,7 +89,10 @@ func (c *Compressor) get(h uint32, si int) int {
 func (c *Compressor) put(h uint32, si int) {
 	h &= htSize - 1
 	c.table[h] = uint16(si)
+	c.inUse[h/32] |= 1 << (h % 32)
 }
+
+func (c *Compressor) reset() { c.inUse = [htSize / 32]uint32{} }
 
 var compressorPool = sync.Pool{New: func() interface{} { return new(Compressor) }}
 
@@ -95,11 +104,8 @@ func CompressBlock(src, dst []byte) (int, error) {
 }
 
 func (c *Compressor) CompressBlock(src, dst []byte) (int, error) {
-	if c.needsReset {
-		// Zero out reused table to avoid non-deterministic output (issue #65).
-		c.table = [htSize]uint16{}
-	}
-	c.needsReset = true // Only false on first call.
+	// Zero out reused table to avoid non-deterministic output (issue #65).
+	c.reset()
 
 	// Return 0, nil only if the destination buffer size is < CompressBlockBound.
 	isNotCompressible := len(dst) < CompressBlockBound(len(src))