Zeroing out the 128kiB hash table takes a considerable amount of time in repeated calls to CompressBlock on small, poorly compressible blocks. Give the hashtable an 8kiB bitmap indicating which entries are in use and zero that out instead.
    
Linux/amd64, Go 1.17:

    name              old speed      new speed      delta
    CompressRandom-8  1.58GB/s ± 2%  1.94GB/s ± 2%  +23.44%  (p=0.000 n=20+19)
    CompressPg1661-8   826MB/s ± 5%   837MB/s ± 2%     ~     (p=0.101 n=19+20)
    CompressDigits-8   151MB/s ± 1%   150MB/s ± 2%   -0.73%  (p=0.012 n=20+20)
    CompressTwain-8    558MB/s ± 2%   559MB/s ± 2%     ~     (p=1.000 n=20+20)
    CompressRand-8    25.1MB/s ± 2%  25.1MB/s ± 3%     ~     (p=1.000 n=20+20)

Go 1.18beta:

    name              old speed      new speed      delta
    CompressRandom-8  1.58GB/s ± 1%  1.95GB/s ± 1%  +23.66%  (p=0.000 n=19+19)
    CompressPg1661-8   765MB/s ± 4%   764MB/s ± 6%     ~     (p=0.678 n=20+20)
    CompressDigits-8   139MB/s ± 5%   137MB/s ± 5%     ~     (p=0.204 n=20+20)
    CompressTwain-8    532MB/s ± 4%   526MB/s ± 6%     ~     (p=0.283 n=19+20)
    CompressRand-8    21.1MB/s ± 4%  21.2MB/s ± 7%     ~     (p=0.635 n=20+20)

(I'm not sure why 1.18 is slower than 1.17 overall.)